♻️ Refactor Search Modal to Page
