+++ src/app/layout.tsx
@@ -2,7 +2,6 @@ import type { Metadata } from 'next'
 import '../styles/index.scss'
 import Providers from './providers'
 import Header from '@/components/common/Header'
-import Search from '@/components/book/Search'
 
 export const metadata: Metadata = {
   title: '라이북러리 | Libookrary',
@@ -20,7 +19,6 @@ export default function RootLayout({
         <Providers>
           <Header />
           {children}
-          <Search />
         </Providers>
       </body>
     </html>
