+++ src/app/search/page.tsx
@@ -1,0 +1,13 @@
+import SearchContent from '@/components/common/Search/SearchContent'
+import SearchHeader from '@/components/common/Search/SearchHeader'
+
+function page() {
+  return (
+    <div className="search">
+      <SearchHeader />
+      <SearchContent />
+    </div>
+  )
+}
+
+export default page
